Modified the Database script to use signed BigInteger. This was done so long can be used in the WebApp. This resulted in modifying existing files to use long instead of BigInteger.
The new files will be used as an interface between the web application and
the database. Files without "DB" in front are Interfaces, those with "DB"
are implemented classes.

DBItemQuery handle queries to the Items table in the
Database.

+++ src/main/java/edu/cpp/cs580/Database/Objects/Interfaces/Item.java
@@ -1,12 +1,10 @@
 package edu.cpp.cs580.Database.Objects.Interfaces;
 
-import java.math.BigInteger;
-
 public interface Item {
-	public BigInteger getItemID();
+	public long getItemID();
 	public String getSystem();
 	public String getTitle();
-	public void setItemID(BigInteger id);
+	public void setItemID(long id);
 	public void setSystem(String sys);
 	public void setTitle(String tle);
 }
